Moving the 2FA to be accessible for SettingsRoute
