+++ src/v2/Apps/Settings/Routes/EditSettings/SettingsEditSettingsRoute.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { SettingsEditSettingsRoute_me } from "v2/__generated__/SettingsEditSettingsRoute_me.graphql"
 import { createFragmentContainer, graphql } from "react-relay"
 import { SettingsEditSettingsInformationFragmentContainer } from "../EditSettings/Components/SettingsEditSettingsInformation"
-import { SettingsEditSettingsTwoFactorRefetchContainer } from "../EditSettings/Components/SettingsEditSettingsTwoFactor"
+import { SettingsEditSettingsTwoFactorRefetchContainer } from "./Components/SettingsEditSettingsTwoFactor/SettingsEditSettingsTwoFactor"
 import { Column, GridColumns, Join, Separator } from "@artsy/palette"
 
 interface SettingsEditRouteProps {
